backend/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# Redis
redis>=5.0,<6.0
django-redis>=5.4,<6.0

# Document processing
python-docx>=1.1,<2.0
mammoth>=1.6,<2.0
beautifulsoup4>=4.12,<5.0

# LLM clients
openai>=1.0,<2.0
anthropic>=0.18,<1.0
google-generativeai>=0.4,<1.0
httpx>=0.26,<1.0

# Security
python-dotenv>=1.0,<2.0
cryptography>=42.0,<43.0

# Utils
Pillow>=10.0,<11.0

# Development
black>=24.0,<25.0
flake8>=7.0,<8.0
